requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.56.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langextract @ git+https://github.com/google/langextract.git@afc7fd9d2364fad2b724cde697b865082ee9ff32
MarkupSafe==3.0.3
ml_collections==1.1.0
more-itertools==10.8.0
multidict==6.7.0
narwhals==2.14.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMuPDF==1.26.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
yarl==1.22.0
